backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-cors-headers>=4.3
drf-yasg>=1.21
dj-database-url>=2.1
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0.0